Display pretty time when featuring AvalonId
