+++ reveries/lib.py
@@ -4,10 +4,17 @@ import sys
 import math
 import logging
 import contextlib
+import datetime
+import uuid
 import pyblish.util
 import avalon.io
 import avalon.api
 from avalon.vendor import requests
+
+try:
+    import bson
+except ImportError:
+    pass
 
 log = logging.getLogger(__name__)
 
@@ -18,6 +25,15 @@ DEFAULT_MATRIX = [1.0, 0.0, 0.0, 0.0,
                   0.0, 1.0, 0.0, 0.0,
                   0.0, 0.0, 1.0, 0.0,
                   0.0, 0.0, 0.0, 1.0]
+
+
+class LocalTZ(datetime.tzinfo):
+    """Local time zone info, I guess. ¯\_(ツ)_/¯ """
+    delta = datetime.datetime.now() - datetime.datetime.utcnow()
+    utcoffset = dst = lambda self, dt: self.delta
+
+
+localtz = LocalTZ()
 
 
 def matrix_equals(a, b, tolerance=1e-10):
@@ -270,3 +286,15 @@ class pindict(dict):  # For experimental code style
             else:
                 new[key] = value
         return new
+
+
+def avalon_id_timestamp(id):
+    if "-" in id:
+        _ut = uuid.UUID(id + "-0000-000000000000").time
+        stm = (_ut - 0x01b21dd213814000) * 100 / 1e9
+        time = datetime.datetime.fromtimestamp(stm)
+    else:
+        time = bson.ObjectId(id).generation_time
+        time = time.astimezone(localtz)
+
+    return time.strftime("%Y%m%dT%H%M%SZ")
